% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{export_to_json}
\alias{export_to_json}
\title{Export Data to a JSON File}
\usage{
export_to_json(data, file_name = "tiddlers.json")
}
\arguments{
\item{data}{A list containing \code{pages} and \code{blocks} data.tables.}

\item{file_name}{The name of the JSON file to write to. Defaults to "tiddlers.json".}
}
\value{
NULL. The JSON file is written to disk as a side effect.
}
\description{
Converts pages and blocks to a JSON-compatible format and writes the data to a file.
}
\details{
The function processes \code{pages} and \code{blocks} to ensure compatibility with the JSON format.
Column names are modified to match the required format, and the combined data is written to the specified file.
}
