% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{export_to_server}
\alias{export_to_server}
\title{Export Data to a Server}
\usage{
export_to_server(data)
}
\arguments{
\item{data}{A list containing the data to be exported.}
}
\value{
Placeholder return value.
}
\description{
This function is a placeholder for exporting data to a server.
}
