% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{export_to_tids}
\alias{export_to_tids}
\title{Export Data to Tiddler Files}
\usage{
export_to_tids(data)
}
\arguments{
\item{data}{A list containing \code{pages} and \code{blocks} data.tables to be exported.}
}
\value{
Placeholder return value.
}
\description{
This function is a placeholder for exporting data to individual tiddler files.
}
