% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_other_keywords}
\alias{extract_other_keywords}
\title{Extract Custom Keywords from Block Content}
\usage{
extract_other_keywords(blocks, keywords_of_interest, remove_keywords = FALSE)
}
\arguments{
\item{blocks}{A data.table containing blocks with a \code{content} column.}

\item{keywords_of_interest}{A character vector of keywords to search for.}

\item{remove_keywords}{Logical. If TRUE, removes keywords from the \code{content} column after extraction. Defaults to FALSE.}
}
\value{
The updated \code{blocks} data.table with keywords extracted and added to the \code{lq_keywords} column.
}
\description{
Searches for specified keywords in the \code{content} column and adds them to the \code{lq_keywords} column.
}
