% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_tags}
\alias{extract_tags}
\title{Extract Tags from Block Content}
\usage{
extract_tags(blocks)
}
\arguments{
\item{blocks}{A data.table containing blocks with a \code{content} column.}
}
\value{
The updated \code{blocks} data.table with tags extracted and added to the \code{tags} column.
}
\description{
Searches for hashtags in the \code{content} column of a data.table and assigns them to the \code{tags} column.
}
