% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{import_from_http}
\alias{import_from_http}
\title{Import Data from an HTTP Endpoint}
\usage{
import_from_http()
}
\value{
Placeholder return value.
}
\description{
This function is a placeholder for importing data from an HTTP endpoint.
}
