% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{import_from_json}
\alias{import_from_json}
\title{Import Data from a JSON File}
\usage{
import_from_json(file_name)
}
\arguments{
\item{file_name}{The name of the JSON file to import.}
}
\value{
A list containing:
\describe{
\item{pages}{A data.table with extracted pages.}
\item{blocks}{A data.table with extracted blocks.}
}
}
\description{
Reads a JSON file and extracts pages and blocks, converting them to tiddler-compatible formats.
}
