% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{json_block_to_tiddler}
\alias{json_block_to_tiddler}
\title{Convert a JSON Block to Tiddler Format}
\usage{
json_block_to_tiddler(data, page_id, block_id = NA, ancestors = NA)
}
\arguments{
\item{data}{A JSON object representing a block.}

\item{page_id}{The ID of the page the block belongs to.}

\item{block_id}{(Optional) The ID of the parent block.}

\item{ancestors}{(Optional) A vector of ancestor block IDs.}
}
\value{
A data.table containing the block's metadata, content, and properties.
}
\description{
Recursively converts a JSON block and its children into a tiddler-compatible data.table.
}
