% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{json_blocks_to_tiddler}
\alias{json_blocks_to_tiddler}
\title{Convert JSON Blocks to Tiddler Format}
\usage{
json_blocks_to_tiddler(data)
}
\arguments{
\item{data}{A list of JSON objects representing blocks.}
}
\value{
A data.table containing all blocks in tiddler-compatible format.
}
\description{
Converts a list of JSON block entries into a combined data.table.
}
