% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{json_page_to_tiddler}
\alias{json_page_to_tiddler}
\title{Convert a JSON Page to Tiddler Format}
\usage{
json_page_to_tiddler(data)
}
\arguments{
\item{data}{A JSON object representing a page.}
}
\value{
A data.table containing the page's \code{id}, \code{page_name}, and properties (if any).
}
\description{
Converts a single JSON page entry into a tiddler-compatible data.table.
}
