% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{json_pages_to_tiddler}
\alias{json_pages_to_tiddler}
\title{Convert JSON Pages to Tiddler Format}
\usage{
json_pages_to_tiddler(data)
}
\arguments{
\item{data}{A list of JSON objects representing pages.}
}
\value{
A data.table containing all pages in tiddler-compatible format.
}
\description{
Converts a list of JSON page entries into a combined data.table.
}
