% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{process_raw_blocks}
\alias{process_raw_blocks}
\title{Process Raw Blocks Data}
\usage{
process_raw_blocks(data, keywords = NA)
}
\arguments{
\item{data}{A list containing \code{blocks} and \code{pages} data.tables.}
}
\value{
The input data, invisibly modified in place.
}
\description{
Processes raw block data by extracting tags, task-related keywords, and transforming links.
}
