% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tid_collapse}
\alias{tid_collapse}
\title{Collapse Strings into a TiddlyWiki-Compatible Format}
\usage{
tid_collapse(strlist)
}
\arguments{
\item{strlist}{A character vector of strings to be collapsed.}
}
\value{
A single string with each element wrapped in \verb{[[...]]}.
}
\description{
Combines a list of strings into a single string with each element enclosed in double square brackets.
}
