% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_links}
\alias{transform_links}
\title{Transform Links in Block Content}
\usage{
transform_links(blocks, pages)
}
\arguments{
\item{blocks}{A data.table containing blocks with a \code{content} column.}

\item{pages}{A data.table containing \code{page_name} and \code{id} columns for mapping links.}
}
\value{
The updated \code{blocks} data.table with links transformed.
}
\description{
Replaces links to page names in the \code{content} column with links to page IDs.
}
